% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{bin2d}
\alias{bin2d}
\title{bin multidimensional data into a two-way contingency table}
\usage{
bin2d(
  x,
  column1,
  column2,
  WhetherNominal1 = FALSE,
  WhetherNominal2 = FALSE,
  breaks1 = "Sturges",
  breaks2 = "Sturges"
)
}
\arguments{
\item{x}{Data matrix: every column refers to a categorical variable}

\item{column1, column2}{according to the column1-th,column2-th columns to classify}

\item{WhetherNominal1}{logical;if \code{TRUE}, the column1-th variable is a nominal variable;if \code{FALSE},the column1-th variable is a numerical variable}

\item{WhetherNominal2}{logical;if \code{TRUE}, the column2-th variable is a nominal variable;if \code{FALSE},the column2-th variable is a numerical variable}

\item{breaks1, breaks2}{see the argument \code{breaks} of \code{Histograms}}
}
\value{
a list including the table of bin frequencies, vectors of breakpoints, and vectors of midpoints
}
\description{
a function to bin multidimensional data into a two-way contingency table
}
\examples{
\dontrun{
bin2d(iris,1,5,FALSE,TRUE)
}
}
