% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatCompR.R
\name{Dplot}
\alias{Dplot}
\title{construct density polygon of bivariate data}
\usage{
Dplot(x, breaks1 = "Sturges", breaks2 = "Sturges")
}
\arguments{
\item{x}{Data matrix x is n by 2}

\item{breaks1, breaks2}{see the argument \code{breaks} of \code{Histograms}}
}
\value{
a bivariate frequency polygon
}
\description{
construct a bivariate frequency polygon
}
\examples{
\dontrun{
Dplot(faithful)
Dplot(matrix(rnorm(1000*2),1000,2))
}
}
